toPrint MatrixPoly, Poly, bug fix in SubstractPolyToPoly, fix unreference of variables

+++ scilab/modules/elementary_functions/includes/elem_common.h
@@ -46,8 +46,8 @@
 #define getRelativeMachinePrecision()		C2F(dlamch)((char*)"e", 1L)
 #define getUnderflowThreshold()					C2F(dlamch)((char*)"u", 1L)
 #define getOverflowThreshold()					C2F(dlamch)((char*)"o", 1L)
-#define isZero(x)												fabs(x) <= getRelativeMachinePrecision()
-#define isEqual(x,y)										fabs(x) - fabs(y) <= getRelativeMachinePrecision()
+#define isZero(x)												fabs((double)x) <= getRelativeMachinePrecision()
+#define isEqual(x,y)										fabs((double)x) - fabs((double)y) <= getRelativeMachinePrecision()
 
 extern double C2F(dlamch)  (char *_pszCommand, unsigned long int);
 extern double C2F(logp1) (double *_pdblVal);
